Reorganized the oracle and actus modules.
